Make CPAN::Meta::YAML optional
If it's available, it will be used to read a plan file.  If not, the
test plan file will be ignored with a warning.

This also only uses an encoding layer on v5.8 or greater.  That is
probably redundant since CPAN::Meta::YAML requires v5.8.1, but should
that ever change, the code is protected.